angle: version chromium-6503
The strategy for pulling ANGLE sources is now different. We can not use
the default git strategy as it would fail while pulling submodules. So
we get the archived sources, unfortunately those do not build, so end up
unused.

The install script then takes care of cloning the ANGLE repo, applying
some patches required for installing and bottling it correctly.

Signed-off-by: Antonio Caggiano <quic_acaggian@quicinc.com>